// src/web/src/api/generated/model/dashboardDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Zalihe.Web | v1
 * OpenAPI spec version: 1.0.0
 */
import type { RecentMovementDto } from './recentMovementDto';
import type { ReorderItemDto } from './reorderItemDto';

export interface DashboardDto {
  stockValue: number;
  activeItems: number;
  belowMinimum: number;
  outOfStock: number;
  reorder: ReorderItemDto[];
  recentMovements: RecentMovementDto[];
}