// src/web/src/api/generated/model/recentMovementDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Zalihe.Web | v1
 * OpenAPI spec version: 1.0.0
 */
import type { StockMovementSource } from './stockMovementSource';
import type { StockMovementType } from './stockMovementType';

export interface RecentMovementDto {
  id: string;
  itemId: string;
  itemName: string;
  type: StockMovementType;
  quantity: number;
  occurredAt: string;
  source: StockMovementSource;
  /** @nullable */
  externalRef: string | null;
  /** @nullable */
  note: string | null;
}