// src/web/src/api/generated/model/reorderItemDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Zalihe.Web | v1
 * OpenAPI spec version: 1.0.0
 */
import type { StockStatus } from './stockStatus';
import type { Unit } from './unit';

export interface ReorderItemDto {
  itemId: string;
  name: string;
  sku: string;
  unit: Unit;
  stock: number;
  minStock: number;
  status: StockStatus;
}